boxes/views/buttons/core: Add EmojiPickerView to add emojis to message.
A new pop-up EmojiPickerView will be triggered on pressing ':' key on a
message. This view will be used to search and select the emoji that a
user wants to toggle(react/unreact with). The selection will be applied
only after exiting the popup.

A minimal search-box implemented on top of PanelSearchBox is also added,
which can be triggered on pressing 'p' key from Emoji picker view. The
search logic will be added in subsequent commits.

Tests added.

Co-authored-by: Kaustubh Nair <kaustubh.nair108@gmail.com>