buttons/views: Add reaction_count param to EmojiButton class.
A new reaction_count field is added to the EmojiButton class to keep
in check the number of reactions with a particular emoji on a message.
This change makes it easier for displaying the emoji_count beside
name and sorting the list based on this count. This commit just calculates
and stores the count for each emoji, updating the UI will be done in
a later commit.

Tests updated.