buttons/views: Show reaction_count in EmojiPicker view.
The EmojiPicker view now shows the count of reactions of a
particular emoji, beside emoji_name. The emoji_list is also
sorted in descending order of emoji_count, then in ascending
order of emoji_names, if the count is equal.
The search mechanism is also refactored to use the
generate_emoji_buttons function to generate the updated list
of emoji for every new search text.

Tests amended.