boxes: Add support for renaming "(no topic)" messages.
This commit adds support for editing messages with no topic, depending
on whether the server setting "realm_allow_message_editing" is enabled
or not.

Any user can now edit the topic of a topicless message (shown as
"(no topic)") in ZT. This should bring message-editing functionality in
ZT more closer to the behaviour of other zulip clients and the webapp in
general.

Tests amended.

Fixes #837.