model/views: Rename react_to_message() to toggle_message_reaction().
The "react_to_message()" function of model is used to both add/remove
reactions from a message. i.e. to toggle reactions for a message.
"react"-to-message imparts the impression that it only reacts(i.e.
adds a reaction) to a message and doesn't hint much about the opposite.

I think toggle_message_reaction() would be a name that is better suited
to the purpose of this function. Since, this function is not used in a
lot of places currently, it's also better to rename it here itself.

Tests amended.