model: Add function to check if the user has reacted with an emoji.
This commit adds a new function in model "has_user_reacted_to_msg"
that checks if the current logged-in user has reacted to the message
with a particular emoji or not.
This method will be particularly useful in further commits to check
user reaction and appropriately update the UI based on that.

Tests added.