bugfix: model: Update mention flags on edit event for indexed message.
This commit updates the index, unread_count and flags if the message was
indexed when an update_message event with change in mentioned flag(s) is
detected. This currently doesn't handle updates when the event is for a
non-indexed message.

Changes in mentioned narrow after this event are not rendered at the same
time but it updates the index so that a return to the mentioned messages
narrow will do so.

Tests amended.