bugfix: Disable navigation keys when focused in popup search-box.
This commit fixes a bug that allows typing letters mapped to secondary
navigational keys inside popup search-box. We started supporting search
box inside PopupView in the previous commit, that had a bug and prevented
letters that were mapped to navigational commands to not get typed in
the search-box(e.g. j,k,G).

This commit adds a new class variable in PopupView that checks if focus
is inside a Search box and disables all navigational keys allowing
those letters to be typed in the search-box.

Tests amended.