helper/model: Copy unread mentioned messages from initial_data to Index.
This commit adds "unread_mentioned_msg_ids" field in model.index that
copies and stores the message ids of unread mentioned messages obtained
from the initial_data into this field. This field is currently not in
sync with the server, but we only copy this data to index so that we are
able to properly update it through events afterwards.

Tests amended.