core/helper/ui: Replace old platform booleans to one PLATFORM variable.
This commit adapts our codebase to use the new PLATFORM variable instead
of the 3 booleans that stood for each of the supported platforms i.e. WSL,
MACOS and LINUX. This ensures that only one platform (OS) is set for each
instance i.e. it is mutually exclusive.

We modify the platform-dependent codes in various modules to use this new
variable, which prevents bugs that arised earlier due to multiple OS'es
being detected in WSL. Discussion about the bug and how it affected some
features in WSL: #zulip-terminal>Platform handling (#T791) and WSL issues.

Tests amended.
Fixes #791.