buttons/views: Require named parameter for initializing TopicButton.
This commit adds the "*" method of passing required parameters using
keyword arguments to TopicButton. This ensures that we pass the correct
parameters to our class and don't miss any, thus preventing bugs and
reducing connascence between parts of the program.

We also remove the default value of count and width from the parameters,
because we are explicitly calling it with some value, and this will help
avoid any potential issues in the future.