platform_code/helper/model: Move notify() code to platform_code module.
This commit shifts the notify() function to the platform_code.py module
in an effort to move all platform dependent code to a single isolated
module so that other modules are unaware of the underlying platform and
how they are handled for different platforms. This will also allow us to
better handle unsupported platforms.

Tests amended.