refactor: tests: ui_tools: Move and rename paramters of EDIT_MESSAGE.
This renames the "msg_body_edit_enabled" parameter of EDIT_MESSAGE keypress
to an expected type, since its not an input. Also, moves it to the last of
the parameters where we generally put expected_*, to maintain consistency.

We also rename the expect_editing_to_succeed parameter to use "expected_*"
prefix and move it to the last to maintain consistency.