tests: ui_tools: Expand stream/private dependent pairs in EDIT_MESSAGE.
This commit expands the stream/private dependednt pairs,
"expected_msg_body_edit_enabled" and "expected_editing_to_succeed", in
EDIT_MESSAGE keypress to include success/failure for both types of message.

We define both the parameters as Dictionaries with keys for stream and
private, each mapping to the expected value of the parameter for that
message type. This allows us to remove the explicit setting of these
parameters from the test body and instead define them in the parametrize.