api_types: Include "flags" in UpdateMessageEvent api type.
This commit adds "flags" field in UpdateMessageEvent api_type, now
that we are planning to handle flag changes in update_message_events.

In particular, users can add/remove several flags such as mentioned/
wildcard_mentioned/has_alert_words from sent messages, and when they
do, it triggers an update_message_event to the rest of the users,
alongwith change in flags if there are any.