bugfix: model: Update mention flags on edit event for unindexed message.
This commit handles updating message flag for un-indexed messages on edit
message events. If the message is not in index, we just need to update the
unread_count. The count is updated after checking with initial data's
unread mentioned msgs field which does not stay synced with the server,
hence the logic is currently buggy.